+++ src/main/java/org/influxdb/InfluxDBFactory.java
@@ -2,10 +2,10 @@ package org.influxdb;
 
 import org.influxdb.impl.InfluxDBImpl;
 
-import com.google.common.base.Preconditions;
-import com.google.common.base.Strings;
-
 import okhttp3.OkHttpClient;
+import org.influxdb.impl.Preconditions;
+
+import java.util.Objects;
 
 
 /**
@@ -25,7 +25,7 @@ public enum InfluxDBFactory {
    * @return a InfluxDB adapter suitable to access a InfluxDB.
    */
   public static InfluxDB connect(final String url) {
-    Preconditions.checkArgument(!Strings.isNullOrEmpty(url), "The URL may not be null or empty.");
+    Preconditions.checkNonEmptyString(url, "url");
     return new InfluxDBImpl(url, null, null, new OkHttpClient.Builder());
   }
 
@@ -42,8 +42,8 @@ public enum InfluxDBFactory {
    * @return a InfluxDB adapter suitable to access a InfluxDB.
    */
   public static InfluxDB connect(final String url, final String username, final String password) {
-    Preconditions.checkArgument(!Strings.isNullOrEmpty(url), "The URL may not be null or empty.");
-    Preconditions.checkArgument(!Strings.isNullOrEmpty(username), "The username may not be null or empty.");
+    Preconditions.checkNonEmptyString(url, "url");
+    Preconditions.checkNonEmptyString(username, "username");
     return new InfluxDBImpl(url, username, password, new OkHttpClient.Builder());
   }
 
@@ -57,8 +57,8 @@ public enum InfluxDBFactory {
    * @return a InfluxDB adapter suitable to access a InfluxDB.
    */
   public static InfluxDB connect(final String url, final OkHttpClient.Builder client) {
-    Preconditions.checkArgument(!Strings.isNullOrEmpty(url), "The URL may not be null or empty.");
-    Preconditions.checkNotNull(client, "The client may not be null.");
+    Preconditions.checkNonEmptyString(url, "url");
+    Objects.requireNonNull(client, "client");
     return new InfluxDBImpl(url, null, null, client);
   }
 
@@ -78,9 +78,9 @@ public enum InfluxDBFactory {
    */
   public static InfluxDB connect(final String url, final String username, final String password,
       final OkHttpClient.Builder client) {
-    Preconditions.checkArgument(!Strings.isNullOrEmpty(url), "The URL may not be null or empty.");
-    Preconditions.checkArgument(!Strings.isNullOrEmpty(username), "The username may not be null or empty.");
-    Preconditions.checkNotNull(client, "The client may not be null.");
+    Preconditions.checkNonEmptyString(url, "url");
+    Preconditions.checkNonEmptyString(username, "username");
+    Objects.requireNonNull(client, "client");
     return new InfluxDBImpl(url, username, password, client);
   }
 }